Permit overriding built-in templates
Any Form with prefix 'smd_ebook_' will be used and will override the correspondingly named plugin templates. Not sure how useful this is in practical terms, but could be handy if the spec changes or for special effects. Note that it still doesn't allow custom content to be injected via <txp:...> tags due to the way the content is built. Investigation underway.

Also: remove double parse() that simulated secondpass, since it's being phased out and the plugin hack is (probably) not needed any more.